+++ src/document.rs
@@ -202,7 +202,7 @@ fn print_directive(directive: &DirectiveDefinition) -> String {
     .iter()
     .map(|arg| {
       let type_str = format!("{}", arg.node.ty.node);
-      if type_str.starts_with('[') {
+      if type_str.starts_with('[') || type_str.starts_with('{') {
         let parts: Vec<&str> = type_str.split(',').collect();
         format!("{}: {}", arg.node.name.node, parts.join(", "))
       } else {
