+++ tests/graphql/test-batching-group-by.graphql
@@ -1,0 +1,43 @@
+#> server-sdl
+schema @server(baseURL: "http://abc.com", batch: {delay: 1, headers: [], maxSize: 1000}) {
+  query: Query
+}
+
+type Post {
+  body: String
+  id: Int
+  title: String
+  user: User @http(path: "/users", query: [{key: "id", value: "{{value.userId}}"}]) @batch(key: "id", path: ["id"])
+  userId: Int
+}
+
+type Query {
+  posts: [Post] @http(path: "/posts?id=1&id=11")
+}
+
+type User {
+  id: Int
+  name: String
+}
+
+#> client-sdl
+type Post {
+  body: String
+  id: Int
+  title: String
+  user: User
+  userId: Int
+}
+
+type Query {
+  posts: [Post]
+}
+
+type User {
+  id: Int
+  name: String
+}
+
+schema {
+  query: Query
+}
